+++ src/main/java/com/example/forummanagementsystem/repositories/PostRepository.java
@@ -23,7 +23,7 @@ public interface PostRepository {
 
     List<Post> getPostsByUserId(Long userId);
 
-    List<Post> filter(Optional<String> title, Optional<String> content, Optional<Integer> rating, Optional<LocalDateTime> createTime, Optional<LocalDateTime> updateTime, Optional<String> sort);
+    List<Post> filter(Optional<String> title, Optional<String> content, Optional<Long> rating, Optional<LocalDateTime> createTime, Optional<LocalDateTime> updateTime, Optional<String> sort);
 
     void update(Post post, User user);
